Update to new asyncio v3
