+++ lib/mqtt_as.py
@@ -24,7 +24,7 @@ from micropython import const
 gc.collect()
 from sys import platform
 
-VERSION = (0, 5, 0)
+VERSION = (0, 6, 0)
 
 # Default short delay for good SynCom throughput (avoid sleep(0) with SynCom).
 _DEFAULT_MS = const(20)
@@ -70,29 +70,6 @@ def pid_gen():
 def qos_check(qos):
     if not (qos == 0 or qos == 1):
         raise ValueError('Only qos 0 and 1 are supported.')
-
-
-class Lock:
-    def __init__(self):
-        self._locked = False
-
-    async def __aenter__(self):
-        while True:
-            if self._locked:
-                await asyncio.sleep_ms(_DEFAULT_MS)
-            else:
-                self._locked = True
-                break
-
-    async def __aexit__(self, *args):
-        self._locked = False
-        await asyncio.sleep_ms(_DEFAULT_MS)
-
-    def locked(self):
-        return self._locked
-
-    def release(self):
-        self._locked = False
 
 
 # MQTT_base class. Handles MQTT protocol on the basis of a good connection.
@@ -146,7 +123,7 @@ class MQTT_base:
         self.newpid = pid_gen()
         self.rcv_pids = set()  # PUBACK and SUBACK pids awaiting ACK response
         self.last_rx = ticks_ms()  # Time of last communication from broker
-        self.lock = Lock()
+        self.lock = asyncio.Lock()
 
     def _set_last_will(self, topic, msg, retain=False, qos=0):
         qos_check(qos)
